Updated Monogame from 3.5 to 3.8
It compiles. Now to tackle run-time errors

+++ TrueCraft.Client/Rendering/FontRenderer.cs
@@ -116,7 +116,7 @@ namespace TrueCraft.Client.Rendering
                             (int)(glyph.Width * scale),
                             (int)(glyph.Height * scale));
 
-                        spriteBatch.Draw(font.GetTexture(glyph.Page), shadowRectangle, sourceRectangle, new Color(21, 21, 21, alpha));
+                        spriteBatch.Draw(font.GetTexture(glyph.Page), shadowRectangle, sourceRectangle, new Color((byte)21, (byte)21, (byte)21, alpha));
                         spriteBatch.Draw(font.GetTexture(glyph.Page), destRectangle, sourceRectangle, new Color(color, alpha));
                         dx += (int)(glyph.XAdvance * scale);
                     }
